+++ src/services/dispatcher/AvalancheBlockDispatcher.ts
@@ -1,16 +1,19 @@
 import {injectable} from 'inversify';
-import {GasEstimation} from "@umb-network/toolbox/dist/types/GasEstimation";
 import {PayableOverrides} from "@ethersproject/contracts";
 
 import {BlockDispatcher} from './BlockDispatcher';
 import {ChainsIds} from '../../types/ChainsIds';
+import {ChainSubmitArgs} from "../../types/ChainSubmit";
 
 @injectable()
 export class AvalancheBlockDispatcher extends BlockDispatcher {
   readonly chainId = ChainsIds.AVALANCHE;
 
-  protected calculatePayableOverrides(gasMetrics: GasEstimation, nonce?: number): PayableOverrides {
-    this.logger.info('[AvalancheBlockDispatcher] using individual gas settings');
+  protected async calculatePayableOverrides(data: ChainSubmitArgs, nonce?: number): Promise<PayableOverrides> {
+    const gasMetrics = await this.resolveGasMetrics();
+    if (!gasMetrics) return {};
+
+    this.logger.info(`[${this.chainId}] AvalancheBlockDispatcher - using individual gas settings`);
     return {nonce, gasPrice: gasMetrics.gasPrice};
   }
 }