+++ src/types/ChainSubmit.ts
@@ -1,0 +1,9 @@
+export type ChainSubmitArgs = {
+  dataTimestamp: number;
+  root: string;
+  keys: Buffer[];
+  values: Buffer[];
+  v: number[];
+  r: string[];
+  s: string[];
+};